app/backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
pydantic-settings>=2.2.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
celery>=5.3.6
redis>=5.0.3
strawberry-graphql[fastapi]>=0.227.0
alembic>=1.13.1
pytest>=8.1.1
pytest-asyncio>=0.23.6
httpx>=0.27.0
